Added low quality modules check
