+++ Website/src/hooks/useSettings.tsx
@@ -139,6 +139,7 @@ export interface StorageDeclaration {
   accent_scheme: Picker<string, any>;
   eruda_console_enabled: boolean;
   disabled_repos: string[];
+  _low_quality_module: boolean;
   __experimental_local_install: boolean;
   repos: StoredRepo[];
   test: any;
@@ -150,6 +151,7 @@ export const INITIAL_SETTINGS: StorageDeclaration = {
   accent_scheme: accent_colors[0],
   eruda_console_enabled: false,
   disabled_repos: [],
+  _low_quality_module: true,
   __experimental_local_install: false,
   repos: [],
   test: [],
